requirements.txt
asgiref==3.3.2
astroid==2.4.2
Django==3.2.25
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
setuptools==70.0.0
validate-docbr==1.7.0
faker==25.8.0
django_filter==23.5.0